requirements.txt
smolagents>=1.14.0
python-dotenv>=1.0.0
langchain>=0.3.22
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-text-splitters>=0.3.8
unstructured>=0.17.2
python-magic-bin>=0.4.14
pdf2image>=1.16.3
python-docx>=0.8.11
markdown>=3.3.7
pdfminer.six>=20221105
python-magic>=0.4.27
huggingface_hub>=0.22.0
rank_bm25
smolagents[openai]